Move tester package to lib
