+++ pkg/lib/tester/token.go
@@ -1,12 +1,12 @@
-package apptester
+package tester
 
-type AppTesterToken struct {
+type TesterToken struct {
 	TokenID   string `json:"token_id"`
 	ReturnURI string `json:"return_uri"`
 }
 
-func NewTesterToken(returnURI string) *AppTesterToken {
-	return &AppTesterToken{
+func NewTesterToken(returnURI string) *TesterToken {
+	return &TesterToken{
 		TokenID:   newTesterTokenID(),
 		ReturnURI: returnURI,
 	}
